requirements.txt
awsiotsdk>=1.21.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0
